Create  This code contains several parts.
Part 1:
In the first part of the code, the selection_sort function is defined, which implements the selection sorting algorithm. It takes a list of numbers and sorts it in ascending order. The algorithm consists of two nested loops, where at each iteration of the outer loop the smallest element is located and swaps with the element at the current position.
Part 2:
In the second part of the code, a new result list is created, where numbers are sequentially added. The goal is to create a list in which neighboring values will have a difference greater than or equal to 1. For this, a random number generator random.randrange() is used, which selects a new number in a given range, provided that the difference between the new and the previous number is greater than 1.
Part 3:
The third part of the code uses the time module to measure execution time. It contains the tic and toc reference points, and also outputs the operation execution time using a formatted string. After that, the selection_sort function is called to sort the list result.